7-11 clean and error free for now

+++ script.js
@@ -5,21 +5,11 @@ const ITEMS = JSON.parse(localStorage.getItem("items")) || [];
 
 let idNum = localStorage.getItem("idNum") || 0;
 
-// leaving off 7/7. Got some stuff working but it feels like a mess. Also, now when I click on literally anything it is targeted and triggers the deleteItem function?????
-
-//Going to delete alot right now, go to github if I want to get anything back
-
-const LISTTEMPLATE = ITEMS.map((item) => {
-  return `<li id="${item.idNum}"><a href="${item.LINK}" title="${item.DESC}">${
-    item.DESC ? item.DESC : item.LINK
-  }</a><button class="delete-btn">Delete</button></li>`;
-});
-
 function populateItems() {
   LISTING.innerHTML = ITEMS.map((item) => {
-    return `<li id="${item.idNum}"><a href="${item.LINK}" title="${
-      item.DESC
-    }">${
+    return `<li class="list-item" id="${item.idNum}"><a href="${
+      item.LINK
+    }" title="${item.DESC}">${
       item.DESC ? item.DESC : item.LINK
     }</a><button class="delete-btn">Delete</button></li>`;
   }).join("");
@@ -59,29 +49,32 @@ function addItem(e) {
 //Need to delete in two places? The live list and local storage? I figured out how to do the idnumber thing, let's see if that's helpful
 
 function deleteItem() {
-  localStorage.setItem("items", JSON.stringify(ITEMS));
+  let check = confirm("Are you sure?");
+  if (check !== true) {
+    return;
+  } else {
+    let filteredItem = ITEMS.filter((item) => {
+      return item.idNum.toString() === this.parentNode.id;
+    });
 
-  let filteredItem = ITEMS.filter((item) => {
-    return item.idNum.toString() === this.parentNode.id;
-  });
+    console.log(filteredItem[0]);
+    console.log(ITEMS);
+    console.log(ITEMS.indexOf(filteredItem[0]));
 
-  console.log(filteredItem[0]);
-  console.log(ITEMS);
-  console.log(ITEMS.indexOf(filteredItem[0]));
+    let deleteIndex = ITEMS.indexOf(filteredItem[0]);
 
-  let deleteIndex = ITEMS.indexOf(filteredItem[0]);
+    ITEMS.splice(deleteIndex, 1);
 
-  ITEMS.splice(deleteIndex, 1);
+    localStorage.setItem("items", JSON.stringify(ITEMS));
 
-  localStorage.setItem("items", JSON.stringify(ITEMS));
+    populateItems();
 
-  populateItems();
+    const DELETEBTN = document.querySelectorAll(".delete-btn");
 
-  const DELETEBTN = document.querySelectorAll(".delete-btn");
-
-  DELETEBTN.forEach((button) => {
-    button.addEventListener("click", deleteItem);
-  });
+    DELETEBTN.forEach((button) => {
+      button.addEventListener("click", deleteItem);
+    });
+  }
 }
 
 populateItems();
